switching some size_t to int's to prevent some cross-platforms issues (Fedora in this particular case)

+++ opensubdiv/far/bilinearSubdivisionTables.h
@@ -82,7 +82,7 @@ template <class T, class U=T> class FarBilinearSubdivisionTables : public FarSub
 public:
 
     // Memory required to store the indexing tables
-    virtual size_t GetMemoryUsed() const;
+    virtual int GetMemoryUsed() const;
    
     // Compute the positions of refined vertices using the specified kernels
     virtual void Refine( int level, void * data=0 ) const;   
@@ -116,7 +116,7 @@ private:
     typename FarSubdivisionTables<T,U>::template Table<unsigned int>  _F_IT;
 };
 
-template <class T, class U> size_t
+template <class T, class U> int
 FarBilinearSubdivisionTables<T,U>::GetMemoryUsed() const {
     return FarSubdivisionTables<T,U>::GetMemoryUsed()+
         _F_ITa.GetMemoryUsed()+
